third_party/WebKit/LayoutTests/fast/css/nth-child-no-mutations-expected.txt

Not necessary to recalc sibling styles for :nth-child when no mutations have happened.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(testElm, null).color is green
PASS successfullyParsed is true

TEST COMPLETE
